msg: account message enqueue-dequeue and issue time
It is very usefull to track how much time message spent in the
write queue before it was dequeued for issuing to the stack
(socket or any other transport).

Also issuing time to the stack is very important, e.g. how much
time we spend in the sendmsg() syscall in case of sending to
socket.

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>