test/fio,msg: cache MOSDOpReply in a queue on a server side and new completion cb
After this patch all messages are preallocated in a send/recv
queue, which gives ~20k IOPS and ~70 MB/s:

   before: IOPS=114k, BW=444MiB/s, Lat=1.125ms
    after: IOPS=132k, BW=514MiB/s, Lat=0.973ms

Also do not call put() on message completion in order to reduce
atomic ops on a fast path.  Since all messages are prealocated
and controlled by the queue no need to account references.

Beware: not on all code paths msg->complete() is called, has to
        be changed accordingly of course.

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>