msg: chain messages in sl list and send all messages from a queue in one go
This patch reduces number of buffer::ptr::release() and sendmsg() syscalls by
preparing one msghdr for all queued messages without usage of temporary buffers
appending them with each part of a message, but simply filling in msghdr.msg_iov
directly and sending it to kernel in one go.

The queueing is also slightly modified: instead of putting a message to the
vector and then erasing it on dequeue side I simply chain messages in a single
linked list, which does not require any memory allocations and enqueue/dequeue
locks a mutex for a very short period of time.

    before: IOPS=132k, BW=514MiB/s, Lat=0.973ms
     after: IOPS=166k, BW=650MiB/s, Lat=0.769ms

Beware:  I did not consider error handling, priority queues, also I did not
         care about sendmsg(), which can return less bytes and iovec has to
	 be advanced.  This is an experimental perf patch.

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>